# docs/swagger.yaml
basePath: /
definitions:
...
  model.CustomerUpdatePayload:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
...
  model.EmployeeUpdatePayload:
    properties:
      active:
        type: boolean
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  util.Response:
    properties:
      data:
        type: object
      message:
        type: string
    type: object
host: micasvn.ddns.net:9999
info:
  contact: {}
  description: Documentation of Garage Service API
  license: {}
  title: Garage Service API
  version: 1.0.0
paths:
...
  /admin/customers:
    get:
      consumes:
      - application/json
      description: Return a list of customers
      parameters:
      - description: Name keyword
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Customer list
      tags:
      - Admin - Customers
  /admin/customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer
      parameters:
      - description: Customer's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Customer delete
      tags:
      - Admin - Customers
    get:
      consumes:
      - application/json
      description: Return details of a customer
      parameters:
      - description: Customer's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Customer detail
      tags:
      - Admin - Customers
    patch:
      consumes:
      - application/json
      description: Update customer's details
      parameters:
      - description: Customer's ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer Update Payload
        in: body
        name: CustomerUpdatePayload
        required: true
        schema:
          $ref: '#/definitions/model.CustomerUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Customer update
      tags:
      - Admin - Customers
  /admin/employees:
    get:
      consumes:
      - application/json
      description: Returns a list of employees
      parameters:
      - description: Name keyword
        in: query
        name: name
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - List employees
      tags:
      - Admin - Employees
  /admin/employees/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an employee
      parameters:
      - description: Employee's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Employee delete
      tags:
      - Admin - Employees
    get:
      consumes:
      - application/json
      description: Return employee's details
      parameters:
      - description: Employee's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Employee detail
      tags:
      - Admin - Employees
    patch:
      consumes:
      - application/json
      description: Update employee's details
      parameters:
      - description: Employee's ID
        in: path
        name: id
        required: true
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Employee Update Payload
        in: body
        name: EmployeeUpdatePayload
        schema:
          $ref: '#/definitions/model.EmployeeUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Employee update
      tags:
      - Admin - Employees
...
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"